Mov behaviors assembly to a shared project
this makes it easier to consume the assembly by multiple views/targets
