+++ src/_Shared.Native/ChartBehaviour.Android.cs
