+++ src/_Shared.Native/ChartBehaviour.MacCatalyst.cs
