+++ src/_Shared.Native/ChartBehaviour.Windows.cs
