+++ src/_Shared.Native/ChartBehaviour._shared.cs
