+++ src/_Shared.Native/Events/Handler.cs
