+++ src/_Shared.Native/Events/PinchEventArgs.cs
