+++ src/_Shared.Native/Events/PinchHandler.cs
