+++ src/_Shared.Native/Events/ScreenEventArgs.cs
