+++ src/_Shared.Native/Events/ScreenHandler.cs
