+++ src/_Shared.Native/Events/ScrollHandler.cs
