+++ src/_Shared.Native/NativeTicker.Mac.cs
@@ -20,10 +20,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#if WINDOWS
+#if IOS || MACCATALYST
 
+using CoreAnimation;
+using Foundation;
 using LiveChartsCore.Motion;
-using Microsoft.UI.Xaml.Media;
 
 namespace LiveChartsCore.Behaviours;
 
@@ -31,20 +32,27 @@ internal partial class NativeFrameTicker : IFrameTicker
 {
     private IRenderMode _renderMode = null!;
     private CoreMotionCanvas _canvas = null!;
+    private CADisplayLink _displayLink = null!;
 
     public void InitializeTicker(CoreMotionCanvas canvas, IRenderMode renderMode)
     {
         _canvas = canvas;
         _renderMode = renderMode;
+        _displayLink = CADisplayLink.Create(OnFrameTick);
+        _displayLink.AddToRunLoop(NSRunLoop.Main, NSRunLoopMode.Common);
 
         _canvas.Invalidated += OnCoreInvalidated;
-        CompositionTarget.Rendering += OnCompositonTargetRendering;
+
+#if DEBUG
+        System.Diagnostics.Trace.WriteLine(
+            "[LiveCharts Info] FrameSync: CADisplayLink (iOS/Catalyst)");
+#endif
     }
 
     private void OnCoreInvalidated(CoreMotionCanvas obj) =>
         _renderMode.InvalidateRenderer();
 
-    private void OnCompositonTargetRendering(object? sender, object e)
+    private void OnFrameTick()
     {
         if (_canvas.IsValid) return;
         _renderMode.InvalidateRenderer();
@@ -52,11 +60,13 @@ internal partial class NativeFrameTicker : IFrameTicker
 
     public void DisposeTicker()
     {
-        CompositionTarget.Rendering -= OnCompositonTargetRendering;
         _canvas.Invalidated -= OnCoreInvalidated;
 
         _canvas = null!;
         _renderMode = null!;
+        _displayLink.Invalidate();
+        _displayLink.Dispose();
+        _displayLink = null!;
     }
 }
